+++ submitr/submission.py
@@ -42,7 +42,7 @@ from submitr.submission_uploads import (
     lookup_ingestion_submission_from_upload_file
 )
 from submitr.utils import chars, format_path, get_health_page, is_excel_file_name, print_boxed, tobool
-from submitr.validators.structured_data_validator_hook import define_structured_data_validator_hook
+from submitr.validators.utils.structured_data_validator_hook import define_structured_data_validator_hook
 
 
 def set_output_file(output_file):
